refactor database ⚡
