+++ src/database/mongoose.js
@@ -1,0 +1,32 @@
+const mongoose = require("mongoose");
+const { log, success, error } = require("../helpers/logger");
+
+module.exports = {
+  async initializeMongoose() {
+    log(`Connecting to MongoDb...`);
+
+    try {
+      await mongoose.connect(process.env.MONGO_CONNECTION, {
+        keepAlive: true,
+        useNewUrlParser: true,
+        useUnifiedTopology: true,
+        useFindAndModify: false,
+      });
+
+      success("Mongoose: Database connection established");
+    } catch (err) {
+      error("Mongoose: Failed to connect to database");
+      process.exit(1);
+    }
+  },
+
+  schemas: {
+    Giveaways: require("./schemas/Giveaways"),
+    Guild: require("./schemas/Guild").model,
+    Member: require("./schemas/Member").model,
+    Message: require("./schemas/Message").model,
+    ModLog: require("./schemas/ModLog").model,
+    TranslateLog: require("./schemas/TranslateLog").model,
+    User: require("./schemas/User").model,
+  },
+};
