+++ src/database/schemas/Giveaways.js
